rag_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# LangChain - using newer compatible versions
langchain>=0.1.0,<0.2.0
langchain-openai>=0.0.5,<0.1.0
langchain-community>=0.0.20,<0.1.0
langchain-core>=0.1.45,<0.2.0
langsmith>=0.1.0,<0.2.0

openai>=1.12.0
qdrant-client==1.7.0
pinecone-client==3.0.0
pypdf2==3.0.1
python-multipart==0.0.6
pydantic==2.5.2
pydantic-settings==2.1.0
tiktoken==0.5.2
aiohttp==3.9.1

# Install PyTorch CPU-only (lightweight, no CUDA bloat!)
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.3.0+cpu
torchvision==0.18.0+cpu

# Sentence transformers (for FREE HuggingFace embeddings)
sentence-transformers==2.7.0
python-jose[cryptography]==3.3.0
prometheus-client==0.19.0

# FREE LLM Providers
langchain-groq>=0.1.3,<0.2.0

# FREE Embeddings
huggingface-hub>=0.20.2

